requirements.txt
beautifulsoup4>=4
scholarly>=1
tqdm==4.67.0
pymupdf==1.25.0
crossrefapi==1.6.0
requests==2.32.2
